requirements.txt
pytesseract
PyMuPDF
Pillow
streamlit
opencv-python
